refactor to standalone class
