+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/tracking/RequestStatisticsLogger.java
@@ -1,5 +1,6 @@
-package com.linkedin.thirdeye.anomaly.utils;
+package com.linkedin.thirdeye.tracking;
 
+import com.linkedin.thirdeye.anomaly.utils.ThirdeyeMetricsUtil;
 import com.linkedin.thirdeye.api.TimeGranularity;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
@@ -22,8 +23,8 @@ public class RequestStatisticsLogger implements Runnable {
   public void run() {
     try {
       long timestamp = System.nanoTime();
-      RequestStatistics stats = ThirdeyeMetricsUtil.getRequestStatistics(0, timestamp);
-      ThirdeyeMetricsUtil.truncateRequestLog(timestamp);
+      RequestStatistics stats = ThirdeyeMetricsUtil.getRequestLog().getStatistics(0, timestamp);
+      ThirdeyeMetricsUtil.getRequestLog().truncate(timestamp);
 
       RequestStatisticsFormatter formatter = new RequestStatisticsFormatter();
       LOG.info("Recent request performance statistics:\n{}", formatter.format(stats));
